requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Bcrypt==1.0.1
python-dotenv==0.20.0
cryptography==39.0.1
